bot_requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.0
httpx==0.27.0
